+++ src/main/java/frc/robot/commands/SpinConveyorHTimed.java
@@ -11,6 +11,7 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.TestingDashboard;
 import frc.robot.subsystems.BallIntake;
+import frc.robot.subsystems.Conveyor;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
@@ -20,22 +21,22 @@ public class SpinConveyorHTimed extends CommandBase {
    */
 
    Timer m_timer;
-   BallIntake m_ballIntake;
+   Conveyor m_conveyor;
    double m_period;
 
   public SpinConveyorHTimed() {
     // Use addRequirements() here to declare subsystem dependencies.
     
-    addRequirements(BallIntake.getInstance());
+    addRequirements(Conveyor.getInstance());
     m_timer = new Timer();
-    m_ballIntake = BallIntake.getInstance();
+    m_conveyor = Conveyor.getInstance();
 
   }
 
   public static void registerWithTestingDashboard() {
-    BallIntake ballIntake = BallIntake.getInstance();
+    Conveyor conveyor = Conveyor.getInstance();
     SpinConveyorHTimed cmd = new SpinConveyorHTimed();
-    TestingDashboard.getInstance().registerCommand(ballIntake, "Timed", cmd);
+    TestingDashboard.getInstance().registerCommand(conveyor, "Timed", cmd);
   }
 
   // Called when the command is initially scheduled.
@@ -50,14 +51,14 @@ public class SpinConveyorHTimed extends CommandBase {
   public void execute() {
 
     double speed = SmartDashboard.getNumber("Conveyor1MotorSpeed",0.5);
-    m_ballIntake.spinHConveyor(speed);
+    m_conveyor.spinHConveyor(speed);
 
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_ballIntake.spinHConveyor(0);
+    m_conveyor.spinHConveyor(0);
   }
 
   // Returns true when the command should end.
